+++ packages/apps/ability-uniswap-swap/src/lib/ability-checks/index.ts
@@ -1,3 +1,3 @@
 export * from './check-native-token-balance';
-export * from './check-token-in-balance';
+export * from './check-erc20-balance';
 export * from './check-erc20-allowance';