+++ packages/apps/ability-uniswap-swap/src/lib/ability-helpers/index.ts
@@ -2,3 +2,5 @@ export * from './get-pkp-info';
 export * from './get-uniswap-quote';
 export * from './send-uniswap-tx';
 export * from './sign-tx';
+export * from './get-erc20-allowance';
+export * from './get-erc20-contract';