+++ packages/libs/app-sdk/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.2.2 (2025-10-06)
+
+### 🧱 Updated Dependencies
+
+- Updated ability-sdk to 2.3.0
+
 ## 2.2.1 (2025-09-25)
 
 ### 🧱 Updated Dependencies